Comment (nearly) all properties in preperties files
Since a default value is set in the geOrchestra code when any property
is loaded from a properties file, it's not necessary to provide it also
in the datadir properties files, unless it is different from the default.

It also makes it easier for the instance administrator to review their
datadir, focusing only on important properties, and letting the default
values unchanged for the rest of the properties.

See https://github.com/georchestra/datadir/issues/129 for the
motivation.

Note that the default.properties file has been let uncommented, as we
always want these values to be set by the instance administrator.